もろもろの更新

+++ get_data_for_similarity.py
@@ -1,6 +1,7 @@
 
 # -*- coding: utf_8 -*- 
 import sys
+import os
 sys.path.append('/home/okano/lab/recommend/pythonOkn/')
 import basictool
 import urllib2
@@ -19,7 +20,7 @@ br.select_form(nr = 0)  #フォームを入力
 # print br.form
 
 # word_list = load_yml( "word_list.yml"
-f = open("word_list.yml", 'r')
+f = open("word_set/word_list_test.yml", 'r')
 word_list = yaml.load(f)
 f.close() 
 for word in word_list:
@@ -29,18 +30,19 @@ for word in word_list:
 
     br.select_form(nr = 0)  #フォームを入力
     word = word.encode('utf-8')
-    br["search"] = word  #
-    response = br.submit() ;
+    if os.path.exists("data_set/"+word+".txt") == False:  #まだ、検索していなければ
+        br["search"] = word  #
+        response = br.submit() ;
 
-    # print response.geturl()     #飛んだページ
+        # print response.geturl()     #飛んだページ
 
-    # print br.find_link(text=word)  #テストという名前のリンクを探す
-    req =  br.click_link(text=word)     #テストという名前のリンクをクリック
-    br.open(req)
-    # print br.response().read()      #
-    f = open('data_set/'+word+'.txt', 'w')
-    f.write(br.response().read())
-    f.close() 
-    # print br.geturl()           #アドレス表示
+        # print br.find_link(text=word)  #テストという名前のリンクを探す
+        req =  br.click_link(text=word)     #テストという名前のリンクをクリック
+        br.open(req)
+        # print br.response().read()      #
+        f = open('data_set/'+word+'.txt', 'w')
+        f.write(br.response().read())
+        f.close() 
+        # print br.geturl()           #アドレス表示
